move media under io again
