+++ tests/io/_media/test_images.py
@@ -5,7 +5,7 @@ import pathlib
 import pytest
 import torch
 
-from aioway.media import ImageLoader, PillowImageLoader, TvioImageLoader
+from aioway.io._media import ImageLoader, PillowImageLoader, TvioImageLoader
 from aioway.schemas import IsImageTag, extract_tags
